Boot order for the VMs
I notice that if a wrong boot order for the VMs is done, than the ScadaBR will not work properly. I also pointed out this issues in the comment section on your Youtube video.
One solution is to run a specific script in the Chemical Plant VM or shutdown everything and reboot in the right order. Hope this could help other user, it was a big issue for me because I was not able to test the Lab.